Partial completion of ghosting game objects
Creation of additional CreateUnit() accommodating for building and ghosting.
Creation of DestroyUnit()
Ghosting material created and linked to prefab object (One object, 2 material: Ghost and actual)
Creation of an AvailableObject list which will load all object associated to specific factions when game starts (more to do)

ControlManager addition of the use of button press B to spawn a ghost building.

To Do:
- Function for when unit creation is cancelled midway (ie destroy ghost unit with primary selected unit) when any other button is pressed)
- Adding AvailableObjects to faction specifics. (ie load a bunch of models to the "availableObjects" of the chosen faction the user wants to play with)